Move utility variables to class instance in Steam class
The utility variables previously declared at the root level of the Steam class file have now been moved to be instance variables within the Steam class itself. This change regularizes the use of `this` keyword which promotes better code readability and maintenance. Moreover, it helps to avoid potential variable scope clashes and enhances the encapsulation within the Steam class.